+++ src/ops/last.rs
@@ -70,6 +70,9 @@ where
     }
     self.observer.complete();
   }
+
+  #[inline]
+  fn is_stopped(&self) -> bool { self.observer.is_stopped() }
 }
 
 #[cfg(test)]