+++ src/ops/skip_last.rs
@@ -1,6 +1,6 @@
 use crate::prelude::*;
 use observable::observable_proxy_impl;
-use observer::error_proxy_impl;
+use observer::{error_proxy_impl, is_stopped_proxy_impl};
 use std::collections::VecDeque;
 
 #[derive(Clone)]
@@ -68,6 +68,8 @@ where
     }
     self.observer.complete();
   }
+
+  is_stopped_proxy_impl!(observer);
 }
 
 #[cfg(test)]