+++ src/ops/take.rs
@@ -1,4 +1,6 @@
-use crate::observer::{complete_proxy_impl, error_proxy_impl};
+use crate::observer::{
+  complete_proxy_impl, error_proxy_impl, is_stopped_proxy_impl,
+};
 use crate::prelude::*;
 use observable::observable_proxy_impl;
 
@@ -69,6 +71,7 @@ where
   }
   error_proxy_impl!(Err, observer);
   complete_proxy_impl!(observer);
+  is_stopped_proxy_impl!(observer);
 }
 
 #[cfg(test)]