+++ src/ops/take_last.rs
@@ -1,4 +1,4 @@
-use crate::observer::error_proxy_impl;
+use crate::observer::{error_proxy_impl, is_stopped_proxy_impl};
 use crate::prelude::*;
 use observable::observable_proxy_impl;
 use std::collections::VecDeque;
@@ -69,6 +69,8 @@ where
     }
     self.observer.complete();
   }
+
+  is_stopped_proxy_impl!(observer);
 }
 
 #[cfg(test)]